requirements.txt
PyMuPDF>=1.23.0
nltk>=3.8.1
bark>=0.1.5
transformers>=4.31.0
torch>=2.0.0
scipy>=1.11.0
pydub>=0.25.1
numpy>=1.24.0
tqdm>=4.66.0 ## Adds a nice progress bar to the terminal
python-dotenv>=1.0.0